requirements.txt
opencv-python
mediapipe
numpy
pandas
tensorflow
scikit-learn
matplotlib
Flask
flask-cors
Pillow
